extract markdown logic into a class method
